refactor: address PathProvider TODO
